fix: move types to modifier files
