+++ src/modifiers/eventListeners.js
@@ -3,7 +3,12 @@ import type { ModifierArguments, Modifier } from '../types';
 import getWindow from '../dom-utils/getWindow';
 
 // eslint-disable-next-line
-export type Options = {
+export type EventListenersModifier = $Shape<Modifier<any>> & {
+  name: 'eventListeners',
+  options?: $Shape<Options>,
+};
+
+type Options = {
   scroll: boolean,
   resize: boolean,
 };
