+++ src/modifiers/flip.js
@@ -10,7 +10,12 @@ import { bottom, top, start, right, left, auto } from '../enums';
 import getVariation from '../utils/getVariation';
 
 // eslint-disable-next-line
-export type Options = {
+export type FlipModifier = $Shape<Modifier<any>> & {
+  name: 'flip',
+  options?: $Shape<Options>,
+};
+
+type Options = {
   fallbackPlacements: Array<Placement>,
   padding: Padding,
   boundary: Boundary,
