+++ src/modifiers/popperOffsets.js
@@ -2,6 +2,12 @@
 import type { ModifierArguments, Modifier } from '../types';
 import computeOffsets from '../utils/computeOffsets';
 
+// eslint-disable-next-line
+export type PopperOffsetsModifier = $Shape<Modifier<any>> & {
+  name: 'offset',
+  options?: $Shape<{||}>,
+};
+
 function popperOffsets({ state, name }: ModifierArguments<{||}>) {
   // Offsets are the actual position the popper needs to have to be
   // properly positioned near its reference element
